# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MedRise Medical Centre API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List all appointments (admin)
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Book a new appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Appointment booked successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "400":
          description: Invalid input
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get a single appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "404":
          description: Not found
    patch:
      operationId: updateAppointmentStatus
      tags: [appointments]
      summary: Update appointment status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "404":
          description: Not found
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Delete an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredentials"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResult"
        "401":
          description: Invalid credentials
...
  /admin/change-password:
    post:
      operationId: changePassword
      tags: [admin]
      summary: Change own password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success:
                    type: boolean
        "400":
          description: Validation error or wrong current password
        "401":
          description: Not authenticated
  /admin/me:
    get:
      operationId: getAdminMe
      tags: [admin]
      summary: Get current admin session
      responses:
        "200":
          description: Admin info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminInfo"
        "401":
          description: Not authenticated

  # ── CONSULTATIONS (EHR) ──────────────────────────────────────────────────
...
components:
  schemas:
...
    Appointment:
      type: object
      required: [id, patientName, phone, email, service, preferredDate, preferredTime, status, createdAt]
      properties:
        id:
          type: integer
        patientName:
          type: string
        phone:
          type: string
        email:
          type: string
        age:
          type: integer
          nullable: true
        sex:
          type: string
          nullable: true
        service:
          type: string
        preferredDate:
          type: string
        preferredTime:
          type: string
        preferredDoctor:
          type: string
          nullable: true
        message:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        createdAt:
          type: string
    AppointmentInput:
      type: object
      required: [patientName, phone, email, service, preferredDate, preferredTime]
      properties:
        patientName:
          type: string
          minLength: 2
        phone:
          type: string
        email:
          type: string
          format: email
        age:
          type: integer
        sex:
          type: string
        service:
          type: string
        preferredDate:
          type: string
        preferredTime:
          type: string
        preferredDoctor:
          type: string
        message:
          type: string
    AppointmentStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
...
    AdminCredentials:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AdminLoginResult:
      type: object
      required: [success, admin]
      properties:
        success:
          type: boolean
        admin:
          $ref: "#/components/schemas/AdminInfo"
    AdminInfo:
      type: object
      required: [id, username, name]
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        role:
          type: ["string", "null"]
...
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          minLength: 1
        newPassword:
          type: string
          minLength: 6
...